Let compiler inline and optimize MemTable's KeyComparator.
```
$ perf record -e cycles:pp --call-graph dwarf ./db_bench --writes 1000000 --benchmarks="fillrandom" --compression_type none -key_size 256

(...)

Initializing RocksDB Options from the specified file
Initializing RocksDB Options from command-line flags
RocksDB:    version 5.8
Date:       Fri Feb 16 00:12:59 2018
CPU:        8 * Intel(R) Core(TM) i7-6820HQ CPU @ 2.70GHz
CPUCache:   8192 KB
Keys:       256 bytes each
Values:     100 bytes each (50 bytes after compression)
Entries:    1000000
Prefix:    0 bytes
Keys per prefix:    0
RawSize:    339.5 MB (estimated)
FileSize:   291.8 MB (estimated)
Write rate: 0 bytes/second
Read rate: 0 ops/second
Compression: NoCompression
Memtablerep: skip_list
Perf Level: 1
------------------------------------------------
Initializing RocksDB Options from the specified file
Initializing RocksDB Options from command-line flags
DB path: [/tmp/rocksdbtest-1000/dbbench]
fillrandom   :       3.327 micros/op 300584 ops/sec;  102.1 MB/s

(...)

[ perf record: Captured and wrote 143.288 MB perf.data (17815 samples) ]
```

`cycles:pp` (presice PMU's event) distrubution according to `perf report`:
```
-   27,94%     0,06%  db_bench     db_bench             [.] rocksdb::MemTableInserter::PutCF
     27,88% rocksdb::MemTableInserter::PutCF
      - rocksdb::MemTableInserter::PutCFImpl
         - 27,07% rocksdb::MemTable::Add
            - 25,55% rocksdb::InlineSkipList<rocksdb::KeyComparator const&>::Insert<false>
               - 24,44% rocksdb::InlineSkipList<rocksdb::KeyComparator const&>::RecomputeSpliceLevels
                  - rocksdb::InlineSkipList<rocksdb::KeyComparator const&>::KeyIsAfterNode
                     - 2,94% rocksdb::(anonymous namespace)::BytewiseComparatorImpl::Compare
                          __memcmp_sse4_1
                       1,93% rocksdb::GetVarint32PtrFallback
                 0,74% rocksdb::InlineSkipList<rocksdb::KeyComparator const&>::KeyIsAfterNode
            + 0,70% rocksdb::(anonymous namespace)::SkipListRep::Allocate
```
